feat: Implement visual improvements and "pixel-perfect" changes
This commit applies a series of visual and UI/UX enhancements based on your feedback to achieve a more professional and "pixel-perfect" appearance.

Key changes include:

- **License Plate Redesign:** The license plate component has been completely reworked. It now correctly parses and formats Russian-style license plates (`x123xx|123`) and renders them with a realistic, bordered style, including a separate region box and a simulated Russian flag emblem.

- **Icon Overhaul:** All crown icons have been replaced with more contextually appropriate ones:
    - A wrench-and-screwdriver icon (`fa-screwdriver-wrench`) is now used for the master in order lists.
    - A worker icon (`fa-user-cog`) is used in the finance/salary calculation section.
    - The worker icon now correctly turns yellow when a bonus is awarded to the master.

- **Bonus Slider Rework:** The UI for the bonus-awarding slider has been polished with improved CSS for a cleaner, more modern, and easier-to-use feel.
