+++ css/components/buttons.css
@@ -1,0 +1,71 @@
+.btn {
+  padding: 8px 16px;
+  border-radius: 8px;
+  font-weight: 600;
+  border: none;
+  cursor: pointer;
+  transition: var(--transition);
+}
+.btn-accent {
+  background: var(--accent);
+  color: #fff;
+}
+.btn-secondary {
+  background: var(--bg);
+  border: 1px solid var(--border);
+  color: var(--text);
+}
+.btn-danger {
+  background: #ef4444;
+  color: #fff;
+}
+.btn-success {
+  background: #16a34a;
+  color: #fff;
+}
+.btn-full-width {
+  width: 100%;
+  height: 48px;
+}
+.btn-with-icon {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+}
+.btn-icon {
+  background: none;
+  border: none;
+  font-size: 1.2rem;
+  color: var(--text-muted);
+}
+.btn-logout {
+  color: #ef4444;
+}
+.btn-logout:hover {
+  color: #dc2626;
+}
+.btn-biohazard {
+    background-color: #f97316; /* Orange-600 */
+    color: #000000;
+    border: none;
+    width: 80px;
+    height: 80px;
+    border-radius: 12px;
+    cursor: pointer;
+    transition: var(--transition);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+.btn-biohazard .fas {
+    font-size: 3rem;
+}
+.btn-biohazard:hover {
+    background-color: #ea580c; /* Orange-700 */
+    transform: scale(1.05);
+}
+.btn-lg {
+    padding: 12px 24px;
+    font-size: 1.1rem;
+}
