+++ css/components/dashboard.css
@@ -1,0 +1,68 @@
+.dashboard {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 12px;
+  margin-bottom: 16px;
+}
+.dashboard-item {
+    background: var(--panel-bg);
+    border: 1px solid var(--border);
+    border-radius: var(--radius);
+    padding: 12px;
+    display: flex;
+    align-items: center;
+    gap: 12px;
+}
+.dashboard-item-icon .fas {
+  font-size: 22px;
+  color: var(--accent);
+}
+.dashboard-item-title {
+  font-size: 0.8rem;
+  color: var(--text-muted);
+  white-space: nowrap;
+}
+.dashboard-item-value {
+  font-size: 1.3rem;
+  font-weight: 700;
+}
+
+@media (max-width: 768px) {
+    .dashboard { grid-template-columns: repeat(2, 1fr); }
+}
+@media (max-width: 480px) {
+    .dashboard { grid-template-columns: repeat(2, 1fr); }
+}
+@media (max-width: 400px) {
+    #home .dashboard { grid-template-columns: 1fr; }
+}
+
+.finance-header .dashboard {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr); /* Default to 3 columns for desktop */
+    gap: 12px;
+}
+
+@media (max-width: 768px) {
+    /* Override the main dashboard rule for finance specifically */
+    .finance-header .dashboard {
+        grid-template-columns: repeat(2, 1fr); /* 2 columns for tablets */
+    }
+}
+
+@media (max-width: 480px) {
+    .finance-header .dashboard {
+        grid-template-columns: 1fr; /* 1 column for mobile */
+    }
+}
+
+@media (max-width: 375px) {
+    .finance-header .dashboard {
+        grid-template-columns: 1fr;
+    }
+    .finance-header .dashboard-item-value {
+        font-size: 1.2rem;
+    }
+}
+
+.finance-header { margin-bottom: 16px; }
