+++ css/components/forms.css
@@ -1,0 +1,305 @@
+.form-group {
+  margin-bottom: 1.25rem;
+}
+.form-group label {
+  display: block;
+  font-size: 0.9rem;
+  margin-bottom: 0.5rem;
+  color: var(--text-muted);
+}
+.form-group textarea {
+    width: 100%;
+    padding: 12px;
+    background: var(--bg);
+    border: 1px solid var(--border);
+    border-radius: 8px;
+    transition: var(--transition);
+    color: var(--text);
+    font-family: inherit;
+    resize: vertical;
+    min-height: 80px;
+}
+.form-group textarea:focus {
+    border-color: var(--accent);
+    outline: none;
+}
+
+/* Custom Select Component */
+.custom-select-wrapper {
+    position: relative;
+}
+.custom-select {
+    position: relative;
+    height: 48px;
+    background: var(--bg);
+    border: 1px solid var(--border);
+    border-radius: 8px;
+    color: var(--text);
+    cursor: pointer;
+    transition: var(--transition);
+}
+.custom-select:focus-within {
+    border-color: var(--accent);
+}
+.custom-select.disabled {
+    background-color: var(--border);
+    cursor: not-allowed;
+    opacity: 0.7;
+}
+.custom-select-trigger {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 16px;
+    height: 100%;
+}
+.custom-options {
+    position: absolute;
+    top: 105%;
+    left: 0;
+    right: 0;
+    background: var(--panel-bg);
+    border: 1px solid var(--border);
+    border-radius: var(--radius);
+    z-index: 100;
+    max-height: 200px;
+    overflow-y: auto;
+    box-shadow: 0 4px 6px rgba(0,0,0,0.1);
+    opacity: 0;
+    visibility: hidden;
+    transform: translateY(-10px);
+    transition: opacity 0.2s, transform 0.2s, visibility 0.2s;
+}
+.custom-options.active {
+    opacity: 1;
+    visibility: visible;
+    transform: translateY(0);
+}
+.custom-option {
+    padding: 12px 16px;
+    transition: background-color 0.2s;
+    display: block;
+}
+.custom-option:hover {
+    background-color: var(--bg);
+}
+.custom-option.selected {
+    background-color: color-mix(in srgb, var(--accent) 80%, black);
+    color: #fff;
+}
+
+/* Стили для автозаполнения поиска */
+.input-with-icon {
+    position: relative;
+}
+.search-results-list {
+    position: absolute;
+    top: 100%;
+    left: 0;
+    right: 0;
+    background: var(--panel-bg);
+    border: 1px solid var(--border);
+    border-top: none;
+    border-radius: 0 0 var(--radius) var(--radius);
+    z-index: 10;
+    max-height: 200px;
+    overflow-y: auto;
+    box-shadow: 0 4px 6px rgba(0,0,0,0.1);
+    display: none;
+}
+.search-results-list.active {
+    display: block;
+}
+.search-result-item {
+    padding: 10px 16px;
+    cursor: pointer;
+    transition: background-color 0.2s;
+}
+.search-result-item:hover {
+    background-color: var(--bg);
+}
+.search-result-item.disabled {
+    cursor: default;
+    color: var(--text-muted);
+}
+.search-result-item strong {
+    color: var(--text);
+}
+
+/* Улучшения для фильтров архива */
+.client-search-container {
+    display: flex;
+    max-width: 500px;
+    margin: 0 auto;
+}
+.client-search-wrapper {
+    position: relative;
+    flex-grow: 1;
+}
+.client-search-wrapper .form-control {
+    height: 48px;
+    font-size: 1rem;
+    padding-left: 20px;
+    border-radius: 24px 0 0 24px;
+    border-right: none;
+}
+.client-search-wrapper .form-control:focus {
+    z-index: 1;
+}
+#client-search-card .voice-search-btn {
+    border-radius: 0 24px 24px 0;
+    background-color: var(--bg);
+    border: 1px solid var(--border);
+    border-left-width: 0;
+    color: var(--text-muted);
+    font-size: 1.2rem;
+    padding: 0 16px;
+}
+#client-search-card .voice-search-btn:hover {
+    background-color: color-mix(in srgb, var(--text) 5%, var(--bg));
+}
+#client-search-card .voice-search-btn.is-recording {
+    color: var(--accent);
+    border-color: var(--accent);
+    z-index: 2;
+}
+.client-search-wrapper .search-results-list {
+    width: 100%;
+}
+.archive-filters {
+  display: flex;
+  gap: 12px;
+  align-items: center;
+  padding: 16px;
+  border-bottom: 1px solid var(--border);
+  flex-wrap: wrap;
+}
+.archive-filters .form-group {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  margin-bottom: 0;
+}
+.archive-filters .form-control {
+  background: var(--bg);
+  border: 1px solid var(--border);
+  border-radius: 8px;
+  padding: 8px;
+  color: var(--text);
+  max-width: 150px;
+}
+.archive-filters #export-archive-csv {
+  margin-left: auto;
+}
+.order-filters {
+  display: flex;
+  gap: 12px;
+  align-items: center;
+  padding: 12px 16px;
+  border-bottom: 1px solid var(--border);
+}
+.order-filters .form-group {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  margin-bottom: 0;
+}
+.archive-filters .form-group label {
+    font-size: 0.9rem;
+    font-weight: 500;
+    color: var(--text-muted);
+    margin-bottom: 6px;
+    display: inline-flex;
+    align-items: center;
+    gap: 6px;
+}
+.date-range-picker {
+    display: inline-flex;
+    align-items: center;
+    background-color: var(--bg);
+    border: 1px solid var(--border);
+    border-radius: 8px;
+    padding: 0 12px;
+    height: 44px;
+    transition: var(--transition);
+}
+.date-range-picker:focus-within {
+    border-color: var(--accent);
+    box-shadow: 0 0 0 3px color-mix(in srgb, var(--accent) 20%, transparent);
+}
+.date-input-wrapper {
+    position: relative;
+    display: flex;
+    align-items: center;
+}
+.date-range-picker .form-control {
+    background: transparent;
+    border: none;
+    padding: 0;
+    height: 42px;
+    width: 125px;
+    color: var(--text);
+    font-family: inherit;
+    font-size: 1rem;
+}
+.date-range-picker .form-control:focus {
+    outline: none;
+}
+.date-range-picker .picker-icon {
+    color: var(--text-muted);
+    pointer-events: none;
+    position: absolute;
+    right: 0;
+}
+.date-range-separator {
+    color: var(--text-muted);
+    margin: 0 8px;
+}
+.archive-filters .btn-accent {
+    min-width: 160px;
+    justify-content: center;
+}
+.archive-period-buttons {
+    display: flex;
+    gap: 8px;
+}
+/* Стили для кастомного вида date input */
+input[type="date"]::-webkit-calendar-picker-indicator {
+    opacity: 0;
+    position: absolute;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    cursor: pointer;
+}
+
+@media (max-width: 480px) {
+    .archive-filters {
+        flex-direction: column;
+        align-items: stretch;
+        gap: 16px;
+    }
+    .archive-filters .btn, .archive-filters .date-range-picker {
+        width: 100%;
+    }
+    .archive-filters .btn {
+        justify-content: center;
+    }
+    .date-range-picker {
+        flex-direction: column;
+        height: auto;
+        padding: 12px;
+        gap: 8px;
+    }
+    .date-range-picker .form-control {
+        width: 100%;
+    }
+    .date-range-separator {
+        display: none;
+    }
+}
+@media (max-width: 768px) {
+  .archive-filters { flex-direction: column; align-items: stretch; }
+  .archive-filters #export-archive-csv { margin-left: 0; margin-top: 12px; }
+}
