+++ css/components/miscellaneous.css
@@ -1,0 +1,241 @@
+/* 9. Уведомления и прочее */
+#notification-root {
+  position: fixed;
+  top: 20px;
+  right: 20px;
+  z-index: 9999;
+}
+.notification {
+  padding: 12px 16px;
+  border-radius: 8px;
+  color: #fff;
+  box-shadow: 0 5px 15px rgba(0,0,0,0.2);
+  transform: translateX(calc(100% + 20px));
+  transition: transform 0.4s;
+}
+.notification.show {
+  transform: translateX(0);
+}
+.notification.success {
+  background-color: #16a34a;
+}
+.notification.error {
+  background-color: #dc2626;
+}
+.empty-state {
+  padding: 40px;
+  text-align: center;
+  color: var(--text-muted);
+}
+.empty-state p {
+  margin-top: 8px;
+}
+
+.admin-zone-content {
+    padding: 24px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 16px;
+}
+
+@media (max-width: 768px) {
+  .admin-zone-content { grid-template-columns: 1fr; }
+}
+
+.warning-banner {
+  grid-column: 1 / -1;
+  padding: 12px;
+  background-color: color-mix(in srgb, var(--gold) 20%, transparent);
+  color: var(--gold);
+  border: 1px solid var(--gold);
+  border-radius: 8px;
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  margin-bottom: 8px;
+}
+
+/* Quick Actions */
+.quick-actions-grid {
+  padding: 16px;
+  display: grid;
+  gap: 12px;
+  grid-template-columns: 1fr 1fr;
+}
+.quick-actions-grid .btn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+  padding-top: 12px;
+  padding-bottom: 12px;
+  font-size: 1rem;
+}
+.quick-action-bottom {
+    grid-column: 1 / -1;
+    padding-top: 16px;
+    padding-bottom: 16px;
+}
+
+/* Leaderboard */
+.leaderboard-table {
+  width: 100%;
+  border-collapse: collapse;
+}
+.leaderboard-table th, .leaderboard-table td {
+  padding: 12px 16px;
+  text-align: left;
+  border-bottom: 1px solid var(--border);
+}
+.leaderboard-table th {
+  font-size: 0.8rem;
+  color: var(--text-muted);
+}
+.leaderboard-table tr:last-child td {
+  border-bottom: none;
+}
+.leaderboard-place .fas {
+  margin-right: 8px;
+}
+.leaderboard-place .fa-trophy.gold { color: var(--gold); }
+.leaderboard-place .fa-trophy.silver { color: var(--silver); }
+.leaderboard-place .fa-trophy.bronze { color: var(--bronze); }
+
+/* Charts */
+.chart {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+.chart-item {
+  display: grid;
+  grid-template-columns: 100px 1fr auto;
+  gap: 10px;
+  align-items: center;
+}
+.chart-label {
+  text-align: right;
+  font-size: 0.9rem;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+.chart-bar-container {
+  background: var(--bg);
+  border: 1px solid var(--border);
+  border-radius: 4px;
+  height: 24px;
+  overflow: hidden;
+}
+.chart-bar {
+  background-color: var(--accent);
+  height: 100%;
+  border-radius: 4px;
+  transition: width 0.5s ease-in-out;
+}
+.chart-value {
+  font-size: 0.9rem;
+  font-weight: 500;
+  padding-left: 5px;
+}
+
+/* Salary Calculation */
+.salary-calculation-list {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+.salary-item {
+  background: var(--bg);
+  border: 1px solid var(--border);
+  border-radius: 8px;
+  padding: 12px;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+.salary-item-header, .salary-details {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    gap: 16px;
+}
+.salary-item-header {
+  font-weight: 600;
+  margin-bottom: 8px;
+}
+.salary-item-header .master-name {
+  font-size: 1.1rem;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+.salary-item-header .final-salary {
+  font-size: 1.2rem;
+  color: var(--accent);
+}
+.salary-details {
+    font-size: 0.9rem;
+    color: var(--text-muted);
+    padding-top: 8px;
+    border-top: 1px solid var(--border);
+}
+.salary-actions {
+    margin-left: auto;
+    padding-left: 16px;
+}
+.finance-actions {
+  margin-top: 24px;
+}
+
+/* Archived Week Summary */
+.week-summary-list {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  padding: 16px;
+}
+.week-summary-item {
+  background: var(--bg);
+  border: 1px solid var(--border);
+  border-radius: 8px;
+  padding: 12px;
+  cursor: pointer;
+  transition: var(--transition);
+}
+.week-summary-item:hover {
+  border-color: var(--accent);
+}
+.week-summary-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  font-weight: 600;
+  margin-bottom: 4px;
+}
+.week-summary-header .week-revenue {
+  color: var(--accent);
+}
+.week-summary-meta {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  font-size: 0.9rem;
+  color: var(--text-muted);
+}
+
+.list-container .btn-full-width {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 8px;
+}
+
+/* Style for the master icon when a bonus is awarded */
+.salary-item.bonus-awarded .master-icon {
+    color: var(--gold);
+}
