+++ css/components/modal.css
@@ -1,0 +1,167 @@
+/* 8. Модальные окна */
+.modal-backdrop {
+  position: fixed;
+  inset: 0;
+  background-color: rgba(0, 0, 0, 0.7);
+  display: grid;
+  place-items: center;
+  z-index: 2000;
+  padding: 16px;
+  opacity: 0;
+  transition: opacity 0.2s;
+}
+.modal-backdrop.show {
+  opacity: 1;
+}
+.modal-content {
+  background: var(--panel-bg);
+  border-radius: var(--radius);
+  padding: 24px;
+  width: 100%;
+  max-width: 500px;
+  transform: scale(0.95);
+  transition: transform 0.2s;
+  max-height: 90vh;
+  overflow-y: auto;
+}
+.modal-backdrop.show .modal-content {
+  transform: scale(1);
+}
+.modal-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 1.5rem;
+}
+.modal-title {
+  font-size: 1.2rem;
+}
+.modal-close-btn {
+  font-size: 1.5rem;
+  color: var(--text-muted);
+  background: none;
+  border: none;
+}
+.modal-body .form-group {
+  margin-bottom: 1rem;
+}
+.modal-body .form-group:last-child {
+  margin-bottom: 0;
+}
+.form-row {
+    display: flex;
+    gap: 16px;
+}
+.form-row .form-group {
+    flex: 1;
+}
+.modal-footer {
+  display: flex;
+  justify-content: flex-end;
+  gap: 12px;
+  margin-top: 1.5rem;
+}
+.modal-content.modal-xl {
+  max-width: 900px;
+}
+
+/* Стили для модального окна бонусов */
+.bonus-modal .modal-body {
+    display: flex;
+    flex-direction: column;
+    gap: 1.25rem;
+}
+.bonus-info-row {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    font-size: 1rem;
+}
+.bonus-info-row span {
+    color: var(--text-muted);
+}
+.bonus-info-row strong {
+    font-weight: 600;
+}
+.bonus-info-row.total strong {
+    font-size: 1.2rem;
+    color: var(--accent);
+}
+.bonus-modal .form-group {
+    margin-bottom: 0;
+}
+.bonus-modal .form-group label {
+    font-weight: 500;
+    margin-bottom: 1rem;
+}
+.bonus-modal .bonus-percentage-value {
+    color: #16a34a; /* green-600 */
+    font-weight: 700;
+}
+.slider-container {
+    position: relative;
+    height: 24px;
+    display: flex;
+    align-items: center;
+}
+.bonus-slider {
+    -webkit-appearance: none;
+    appearance: none;
+    width: 100%;
+    height: 8px;
+    background: var(--bg);
+    border: 1px solid var(--border);
+    border-radius: 4px;
+    outline: none;
+    padding: 0;
+    margin: 0;
+    cursor: pointer;
+    transition: var(--transition);
+}
+.bonus-slider:hover {
+    border-color: var(--accent);
+}
+.bonus-slider::-webkit-slider-thumb {
+    -webkit-appearance: none;
+    appearance: none;
+    width: 24px;
+    height: 24px;
+    background: #fff;
+    border-radius: 50%;
+    border: 2px solid var(--accent);
+    box-shadow: 0 1px 3px rgba(0,0,0,0.2);
+    transition: transform 0.2s ease;
+}
+.bonus-slider:active::-webkit-slider-thumb {
+    transform: scale(1.15);
+    box-shadow: 0 2px 6px rgba(0,0,0,0.2);
+}
+.bonus-slider::-moz-range-thumb {
+    width: 20px;
+    height: 20px;
+    background: #fff;
+    border-radius: 50%;
+    border: 2px solid var(--accent);
+    box-shadow: 0 1px 3px rgba(0,0,0,0.2);
+}
+.slider-ticks {
+    position: absolute;
+    top: 50%;
+    left: 12px; /* half of thumb width */
+    right: 12px;
+    height: 8px;
+    transform: translateY(-50%);
+    display: flex;
+    justify-content: space-between;
+    pointer-events: none;
+}
+.slider-ticks span {
+    width: 2px;
+    height: 8px;
+    background-color: var(--border);
+    border-radius: 1px;
+}
+
+@media (max-width: 768px) {
+  .modal-content { width: 95%; }
+}
