+++ css/components/theme-switch.css
@@ -1,0 +1,44 @@
+/* 7. Переключатель темы (анимированная шестеренка) */
+.theme-switch {
+  position: relative;
+  display: inline-block;
+  width: 50px;
+  height: 26px;
+}
+.theme-switch input {
+  opacity: 0;
+  width: 0;
+  height: 0;
+}
+.wheel-slider {
+  position: absolute;
+  inset: 0;
+  cursor: pointer;
+  background: var(--bg);
+  border-radius: 26px;
+  border: 1px solid var(--border);
+}
+.wheel-icon-wrapper {
+  position: absolute;
+  height: 20px;
+  width: 20px;
+  left: 3px;
+  bottom: 3px;
+  background: var(--panel-bg);
+  border-radius: 50%;
+  transition: var(--transition);
+  display: grid;
+  place-items: center;
+  box-shadow: 0 1px 3px rgba(0,0,0,0.2);
+}
+.wheel-icon-wrapper .fas {
+  color: var(--accent);
+  font-size: 14px;
+  transition: transform 0.4s ease;
+}
+input:checked + .wheel-slider .wheel-icon-wrapper {
+  transform: translateX(24px);
+}
+input:checked + .wheel-slider .wheel-icon-wrapper .fas {
+  transform: rotate(180deg);
+}
