+++ css/layout.css
@@ -1,0 +1,121 @@
+/* 5. Основная структура приложения */
+.page-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  height: var(--header-height);
+  padding: 0 16px;
+  background: color-mix(in srgb, var(--panel-bg) 85%, transparent);
+  backdrop-filter: blur(10px);
+  -webkit-backdrop-filter: blur(10px);
+  border-bottom: 1px solid var(--border);
+  z-index: 1000;
+  gap: 16px;
+}
+.header-controls {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+}
+.app-title {
+  font-size: 1.2rem;
+  font-weight: 700;
+  color: var(--accent);
+}
+.app-title .title-auto {
+  color: var(--text);
+}
+#user-name-display {
+  font-weight: 500;
+}
+
+.main-content {
+  padding: 16px;
+  max-width: 1200px;
+  margin: 0 auto;
+}
+.tab-content {
+  display: none;
+}
+.tab-content.active {
+  display: block;
+  animation: fadeIn 0.3s ease-in-out;
+}
+@keyframes fadeIn {
+  from { opacity: 0; } to { opacity: 1; }
+}
+
+.nav-container {
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  height: var(--nav-height);
+  z-index: 1000;
+  background: color-mix(in srgb, var(--panel-bg) 85%, transparent);
+  backdrop-filter: blur(10px);
+  -webkit-backdrop-filter: blur(10px);
+  border-top: 1px solid var(--border);
+}
+.nav-tabs {
+  display: flex;
+  justify-content: space-around;
+  height: 100%;
+}
+.nav-tab {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  flex-grow: 1;
+  gap: 4px;
+  color: var(--text-muted);
+  border: none;
+  background: none;
+  transition: var(--transition);
+}
+.nav-tab.active { color: #166534; } /* Dark Green for Home */
+.nav-tab[data-tab="finance"].active { color: var(--gold); }
+.nav-tab[data-tab="archive"].active { color: #ffffff; text-shadow: 0 0 3px rgba(0,0,0,0.4); }
+.nav-tab[data-tab="orders"].active { color: #7e22ce; } /* Purple */
+
+/* Light theme overrides for active nav icons */
+html:not([data-theme="dark"]) .nav-tab[data-tab="finance"].active {
+    color: var(--bronze); /* Dark Yellow */
+}
+html:not([data-theme="dark"]) .nav-tab[data-tab="archive"].active {
+    color: #475569; /* Dark Slate Gray */
+    text-shadow: none;
+}
+
+.nav-tab i { font-size: 1.5rem; }
+.nav-tab span { font-size: 0.7rem; }
+
+.section {
+  background: var(--panel-bg);
+  border-radius: var(--radius);
+  margin-bottom: 16px;
+  border: 1px solid var(--border);
+}
+.section-header {
+  padding: 12px 16px;
+  border-bottom: 1px solid var(--border);
+  font-weight: 600;
+}
+
+.privileged-only { display: none !important; }
+
+/* Show privileged elements correctly based on their type */
+body.is-privileged .nav-tab.privileged-only,
+body.is-privileged .quick-actions-grid .btn.privileged-only {
+    display: flex !important;
+}
+
+body.is-privileged .section.privileged-only,
+body.is-privileged .tab-content.privileged-only.active {
+    display: block !important;
+}
