+++ css/pages/login.css
@@ -1,0 +1,51 @@
+/* 4. Стили страницы входа */
+.login-body {
+  padding: 16px;
+  display: flex; /* Изменено для выравнивания */
+  flex-direction: column;
+  justify-content: center; /* Центрирование по вертикали */
+  align-items: center;
+  height: 100vh;
+  overflow: hidden;
+}
+.login-box {
+  width: 100%;
+  max-width: 380px;
+  padding: 2.5rem;
+  background-color: var(--panel-bg);
+  border-radius: var(--radius);
+  border: 1px solid var(--border);
+  margin-top: -10vh; /* Сдвигаем блок вверх */
+}
+.title-container { display: flex; justify-content: space-between; align-items: center; margin-bottom: 0.5rem; }
+.login-title { font-size: 2rem; font-weight: 700; }
+.login-title .title-auto { color: var(--text-muted); }
+.login-subtitle { text-align: center; color: var(--text-muted); margin-bottom: 2rem; }
+
+.input-wrapper { position: relative; }
+.input-icon { position: absolute; top: 50%; transform: translateY(-50%); left: 15px; color: var(--text-muted); }
+.form-group input { width: 100%; height: 48px; padding-left: 45px; padding-right: 45px; background: var(--bg); border: 1px solid var(--border); border-radius: 8px; transition: var(--transition); color: var(--text); }
+.form-group input:focus { border-color: var(--accent); }
+.toggle-password { position: absolute; top: 50%; transform: translateY(-50%); right: 15px; cursor: pointer; color: var(--text-muted); }
+.btn-login {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.75rem;
+  width: 100%;
+  padding: 14px;
+  font-size: 1.1rem;
+  font-weight: 600;
+  color: #fff;
+  background: var(--accent);
+  border-radius: 8px;
+  border: none;
+  height: 52px;
+  cursor: pointer;
+  transition: var(--transition);
+}
+.btn-login:hover, .btn-login:focus {
+  filter: brightness(1.1);
+}
+.general-error-message { color: #F56C6C; text-align: center; margin-top: 1rem; min-height: 1.2em; }
+.page-footer { position: fixed; bottom: 1rem; left: 50%; transform: translateX(-50%); font-size: 0.8rem; color: var(--text-muted); }
