+++ index.html
@@ -150,6 +150,6 @@
   <div id="notification-root"></div>
 
   <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/4.7.2/socket.io.js"></script>
-  <script src="script.js" defer></script>
+  <script src="js/modules/app.js" type="module" defer></script>
 </body>
 </html>
