+++ server.log
@@ -1,0 +1,4 @@
+[DB] Файл db.json не найден или поврежден. Создаем новую базу.
+[SEED] Запуск генерации тестовых данных...
+[SEED] Создано 50 тестовых заказ-нарядов и 5 клиентов.
+>>> Сервер VIPавто v9.0 запущен на порту 3000 <<<
